pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "secure-rag-pipeline"
version = "0.1.0"
description = "Enterprise Secure RAG Pipeline (FastAPI) with tenant isolation and security guardrails."
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [{ name = "Wellington de Freitas" }]
keywords = ["rag", "security", "fastapi", "llm", "owasp", "tenant-isolation"]

dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.0",
  "pyjwt>=2.8.0",
  "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.5.0",
  "pip-audit>=2.7.0",
  "bandit>=1.7.9",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["."]

[tool.setuptools]
packages = ["app"]
